refactor(ui): consolidate ui configuration
We consolidate all UI configuration in `ui/src/config.ts`. This has the
following advantages

* Configuration is stateless. This removes the need for stores for the
  proxy address and makes the dependency graph simpler.
* We separate Electron IPC communication from configuration.
