+++ ui/App.svelte
@@ -2,7 +2,7 @@
   import Router, { push, location } from "svelte-spa-router";
 
   import * as hotkeys from "./src/hotkeys.ts";
-  import { isExperimental } from "./src/ipc";
+  import { isExperimental } from "./src/config";
   import "./src/localPeer.ts";
   import * as path from "./src/path.ts";
   import * as remote from "./src/remote.ts";
@@ -122,7 +122,7 @@
 <NotificationFaucet />
 <Theme />
 
-{#if isExperimental() && sessionIsUnsealed && $location !== path.designSystemGuide()}
+{#if isExperimental && sessionIsUnsealed && $location !== path.designSystemGuide()}
   <TransactionCenter />
 {/if}
