+++ ui/Hotkeys.svelte
@@ -6,7 +6,7 @@
   import * as screen from "./src/screen";
   import { isMac } from "./src/settings";
   import * as hotkeys from "./src/hotkeys";
-  import { isDev } from "./src/ipc";
+  import { isDev } from "./src/config";
 
   const toggle = (destination: string) => {
     if (path.active(destination, $location)) {
@@ -46,7 +46,7 @@
 
     const shortcut = [
       ...hotkeys.shortcuts,
-      ...(isDev() ? hotkeys.devShortcuts : []),
+      ...(isDev ? hotkeys.devShortcuts : []),
     ].find(shortcut => {
       const match = shortcut.key === event.key;
       return shortcut.modifierKey ? match && modifierKey : match;
