+++ ui/src/api.ts
@@ -1,7 +1,6 @@
 import qs from "qs";
-import { get as getStore, writable } from "svelte/store";
 
-export const backendAddressStore = writable("localhost:17246");
+import * as config from "./config";
 
 interface Options {
   query?: Record<string, unknown>;
@@ -39,7 +38,7 @@ const request = (endpoint: string, init?: Init): Request => {
     endpoint = `${endpoint}?${qs.stringify(init.query)}`;
   }
 
-  return new Request(`http://${getStore(backendAddressStore)}/v1/${endpoint}`, {
+  return new Request(`http://${config.proxyAddress}/v1/${endpoint}`, {
     headers: {
       "Content-Type": "application/json",
     },
