+++ ui/src/ipc.ts
@@ -1,17 +1,8 @@
-// eslint-disable-next-line @typescript-eslint/triple-slash-reference,spaced-comment
-/// <reference path="../../native/preload.d.ts" />
+import type {} from "../../native/preload";
 import * as ipcTypes from "../../native/ipc-types";
+import * as config from "./config";
 
 export type { ProxyError } from "../../native/ipc-types";
-
-// `true` if we are running unit tests with Jest.
-const isNodeTestEnv = Boolean(
-  globalThis.process && globalThis.process.env["NODE_ENV"] === "test"
-);
-
-// `true` if this code is run by the Cypress test driver.
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const isCypressTestEnv = Boolean((globalThis as any).cy);
 
 function makeMainProcessClient(): ipcTypes.MainProcess {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -37,24 +28,11 @@ export const openUrl = mainProcess.openUrl;
 
 export const getGitGlobalDefaultBranch = mainProcess.getGitGlobalDefaultBranch;
 
-// Informs whether it's running in a development environment.
-export const isDev = (): boolean => {
-  return window.electron.isDev;
-};
-
-// Informs whether it's running in experimental mode, where
-// features under construction are enabled and can thus be used.
-// This option can only be enabled iff `isDev()` as we should only
-// want to toggle it while in development mode.
-export const isExperimental = (): boolean => {
-  return window.electron.isExperimental;
-};
-
 // Register a listener for the `ipcTypes.ProxyError` message.
 export function listenProxyError(
   f: (proxyError: ipcTypes.ProxyError) => void
 ): void {
-  if (isNodeTestEnv || isCypressTestEnv) {
+  if (config.isNodeTestEnv || config.isCypressTestEnv) {
     return;
   }
